<test> ChessGameTest로 ChessGame 테스트 코드 수정
